requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
celery==5.5.3
redis==6.4.0
psycopg[binary]==3.2.11
dj-database-url==3.0.1
gunicorn==23.0.0
whitenoise==6.11.0
